requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
--find-links wheels

amqp==5.0.2               # via kombu
aniso8601==7.0.0          # via graphene
anyascii==0.1.7           # via wagtail
argon2-cffi==20.1.0
asgiref==3.3.1            # via django
attrs==20.3.0             # via pytest
beautifulsoup4==4.8.2     # via wagtail
billiard==3.6.3.0         # via celery
boto3==1.16.25
botocore==1.19.37         # via boto3, s3transfer
cabins-app==0.0.15
celery==5.0.0
certifi==2020.12.5        # via requests, sentry-sdk
cffi==1.14.4              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.1.2              # via celery, click-didyoumean, click-repl, prequ
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
coverage==5.3
dj-database-url==0.5.0
django==3.1.3
django-appconf==1.0.4
django-cors-headers==3.2.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0      # via wagtail
django-jinja==2.7.0
django-modelcluster==5.1  # via wagtail
django-redis==4.12.1
django-seo-js==0.3.3
django-storages==1.10.1
django-taggit==1.3.0      # via wagtail
django-treebeard==4.3.1   # via wagtail
djangorestframework==3.12.2  # via wagtail
draftjs-exporter==2.1.7   # via wagtail
et-xmlfile==1.0.1         # via openpyxl
flake8==3.8.4
graphene==2.1.8           # via graphene-django
graphene-django==2.13.0
graphql-core==2.3.2       # via graphene, graphene-django, graphql-relay
graphql-relay==2.0.1      # via graphene
gunicorn==20.0.4
html5lib==1.1             # via wagtail
idna==2.10                # via requests
importlib-metadata==3.3.0  # via flake8, kombu, pluggy, pytest
iniconfig==1.1.1          # via pytest
isort==5.6.4
jdcal==1.4.1              # via openpyxl
jinja2==2.11.2            # via django-jinja
jmespath==0.10.0          # via boto3, botocore
kombu==5.0.2              # via celery
l18n==2020.6.1            # via wagtail
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mock==4.0.3
openpyxl==3.0.5           # via tablib
packaging==20.8           # via pytest
pillow==8.0.1             # via wagtail
pluggy==0.13.1            # via pytest
git+https://github.com/wgarlock/prelaunch#egg=prequ==1.4.7.post0.dev82+gf91a039
promise==2.3              # via graphene-django, graphql-core, graphql-relay
prompt-toolkit==3.0.8     # via click-repl
psycopg2-binary==2.8.5
py==1.10.0                # via pytest
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest==6.2.0
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.1    # via botocore
pytz==2020.4              # via celery, django, django-modelcluster, l18n
pyyaml==5.3.1
redis==3.5.3
requests==2.25.0          # via django-seo-js, wagtail
rx==1.6.1                 # via graphql-core
s3transfer==0.3.3         # via boto3
sentry-sdk==0.19.4
singledispatch==3.4.0.3   # via graphene-django
six==1.15.0               # via argon2-cffi, click-repl, graphene, graphene-django, graphql-core, graphql-relay, html5lib, l18n, promise, python-dateutil, singledispatch
soupsieve==2.1            # via beautifulsoup4
sqlparse==0.4.1           # via django, django-debug-toolbar
tablib[xls,xlsx]==3.0.0   # via wagtail
toml==0.10.2              # via pytest
typing-extensions==3.7.4.3  # via importlib-metadata
unidecode==1.1.1          # via graphene-django, wagtail
urllib3==1.26.2           # via botocore, requests, sentry-sdk
vine==5.0.0               # via amqp, celery
wagtail==2.11.2
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
wheel==0.36.2
willow==1.4               # via wagtail
xlrd==2.0.1               # via tablib
xlsxwriter==1.3.7         # via wagtail
xlwt==1.3.0               # via tablib
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip                       # via prequ
# setuptools                # via gunicorn
